feat: use svgr for converting svgs to JSX
Currently the contents of the svgs are not converted to JSX, which leads to warnings like 'Warning:
Invalid DOM property `clip-rule`. Did you mean `clipRule`?'.

BREAKING CHANGE: This shouldn't really be a breaking change, but it's a fundamental change to how
the icon components are created. No too likely, but it's possible that something is relying on the
old behavior.